requirements.txt
torch
torchvision
Pillow
transformers
pandas
numpy
geopy
flask
